Remove libc
Not everything could be remove without breaking the existing API. In particular, `ts_tree_print_dot_graph` function takes a `FILE *` as input parameter.

This function has been moved to a separate file called `capi.rs`, which is compiled when the `capi` feature is available. Internally the project uses a `BufWriter<File>` normally, but when `capi` feature is requested, a `CFile` is stored. This is an abstraction layer between `FILE *` and `io::Write`, allowing the usage of C stream functions in a transparent way.